+++ src/parent_process.c
@@ -1,0 +1,26 @@
+#include "minishell.h"
+
+void	parent_process(t_exe *exe, pid_t pid, int i)
+{
+	int	status;
+
+	if (i == 0 && exe->pip_cnt == 0)
+	{
+	}
+	else if (i % 2 == 0 && exe->pip_cnt == 0)
+		close(exe->b[READ]);
+	else if (i % 2 != 0 && exe->pip_cnt == 0)
+		close(exe->a[READ]);
+	else if (i % 2 == 0 && exe->pip_cnt > 0)
+	{
+		close(exe->a[WRITE]);
+		if (exe->flag_b != 0)
+			close(exe->b[READ]);
+	}
+	else if (i % 2 != 0 && exe->pip_cnt > 0)
+	{
+		close(exe->a[READ]);
+		close(exe->b[WRITE]);
+	}
+	waitpid(pid, &status, 0);
+}
